Move meta images to separate folder
