+++ src/images/meta/favicon-1200x1200.png
